Zebraix: Reorg and consolidate graph realization.
